requirements.txt
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
google-generativeai==0.3.2
newsapi-python==0.2.7
python-dotenv==1.0.0
mlxtend==0.22.0
requests==2.31.0 